+++ allennlp/models/__init__.py
@@ -27,3 +27,4 @@ from allennlp.models.esim import ESIM
 from allennlp.models.bimpm import BiMpm
 from allennlp.models.graph_parser import GraphParser
 from allennlp.models.bidirectional_lm import BidirectionalLanguageModel
+from allennlp.models.language_model import LanguageModel
