+++ allennlp/modules/token_embedders/bidirectional_language_model_token_embedder.py
@@ -1,21 +1,11 @@
-import json
-from typing import Dict, Tuple, TYPE_CHECKING
+from typing import Tuple
 
-import torch
-
-from allennlp.common.checks import ConfigurationError
-from allennlp.data import TokenIndexer, Token
-from allennlp.modules.scalar_mix import ScalarMix
 from allennlp.modules.token_embedders.token_embedder import TokenEmbedder
-from allennlp.nn.util import remove_sentence_boundaries, get_text_field_mask, add_sentence_boundary_token_ids
-
-# Importing at runtime results in a circular dependency.
-if TYPE_CHECKING:
-    from allennlp.models.bidirectional_lm import BidirectionalLanguageModel
+from allennlp.modules.token_embedders.language_model_token_embedder import LanguageModelTokenEmbedder
 
 
 @TokenEmbedder.register('bidirectional_lm_token_embedder')
-class BidirectionalLanguageModelTokenEmbedder(TokenEmbedder):
+class BidirectionalLanguageModelTokenEmbedder(LanguageModelTokenEmbedder):
     """
     Compute a single layer of representations from a bidirectional language model. This is done
     by computing a learned scalar average of the layers from the LM. Typically the LM's weights
@@ -52,139 +42,8 @@ class BidirectionalLanguageModelTokenEmbedder(TokenEmbedder):
                  bos_eos_tokens: Tuple[str, str] = ("<S>", "</S>"),
                  remove_bos_eos: bool = True,
                  requires_grad: bool = False) -> None:
-        super().__init__()
-
-        overrides = {
-                "model": {
-                        "contextualizer": {
-                                "return_all_layers": True
-                        }
-                }
-        }
-
-        # Import here to avoid circular dependency.
-        from allennlp.models.archival import load_archive
-        # Load LM and the associated config.
-        archive = load_archive(archive_file, overrides=json.dumps(overrides))
-        self._lm: BidirectionalLanguageModel = archive.model
-        self._lm.delete_softmax()
-        config = archive.config
-        dict_config = config.as_dict(quiet=True)
-
-        # Extract the name of the tokens that the LM was trained on.
-        text_field_embedder = dict_config["model"]["text_field_embedder"]
-        token_names = list(text_field_embedder["token_embedders"].keys())
-        if len(token_names) != 1:
-            # We don't currently support embedding with language models trained with multiple
-            # embedded indices.
-            #
-            # Note: We only care about embedded indices. This does not include "tokens" which
-            # is just used to compute the loss in BidirectionalLanguageModel.
-            raise ConfigurationError(f"LM from {archive_file} trained with multiple embedders!")
-        if "embedder_to_indexer_map" in text_field_embedder:
-            # Similarly we don't support multiple indexers per embedder.
-            raise ConfigurationError(f"LM from {archive_file} trained with embedder_to_indexer_map!")
-        self._token_name = token_names[0]
-
-        # TODO(brendanr): Find a way to remove this hack. The issue fundamentally is that the
-        # BasicTextFieldEmbedder concatenates multiple embedded representations. When a
-        # downstream model uses both, tokens and token characters, say, and only adds bos/eos
-        # tokens to the token characters, the dimensions don't match. See:
-        # https://github.com/allenai/allennlp/blob/eff25a3085aa9976a7650d30d8961c3626ddc411/allennlp/modules/text_field_embedders/basic_text_field_embedder.py#L109
-        #
-        # For the equivalent hack in the ELMo embedder see:
-        # https://github.com/allenai/allennlp/blob/eff25a3085aa9976a7650d30d8961c3626ddc411/allennlp/modules/elmo.py#L590
-        if bos_eos_tokens:
-            dataset_reader_config = config.get("dataset_reader")
-            if dataset_reader_config.get("type") == "multiprocess":
-                dataset_reader_config = dataset_reader_config.get("base_reader")
-            token_indexer_config = dataset_reader_config.get("token_indexers").get(self._token_name)
-            token_indexer: TokenIndexer = TokenIndexer.from_params(token_indexer_config)
-            token_list = [Token(token) for token in bos_eos_tokens]
-            # TODO(brendanr): Obtain these indices from the vocab once the
-            # ELMoTokenCharactersIndexer adds the mappings.
-            bos_eos_indices = token_indexer.tokens_to_indices(token_list, self._lm.vocab, "key")["key"]
-            self._bos_indices = torch.Tensor(bos_eos_indices[0])
-            self._eos_indices = torch.Tensor(bos_eos_indices[1])
-        else:
-            self._bos_indices = None
-            self._eos_indices = None
-
-        if dropout:
-            self._dropout = torch.nn.Dropout(dropout)
-        else:
-            self._dropout = lambda x: x
-
-        self._remove_bos_eos = remove_bos_eos
-        num_layers = self._lm.num_layers()
-        # TODO(brendanr): Consider passing our LM as a custom module to `Elmo` instead.
-        # See https://github.com/allenai/allennlp/blob/master/allennlp/modules/elmo.py#L76
-        self._scalar_mix = ScalarMix(mixture_size=num_layers, do_layer_norm=False, trainable=True)
-
-        # pylint: disable=protected-access
-        character_dim = self._lm._text_field_embedder.get_output_dim()
-        contextual_dim = self._lm._contextualizer.get_output_dim()
-
-        if contextual_dim % character_dim != 0:
-            raise ConfigurationError(
-                    "The output dimensions for the text_field_embedder " +
-                    f"({character_dim}) and the contextualizer ({contextual_dim})" +
-                    f" from the language model loaded from {archive_file} are " +
-                    "not compatible. Please check the config used to train that " +
-                    "model and ensure that the output dimension of the " +
-                    "text_field_embedder divides the output dimension of the " +
-                    "contextualizer.")
-        self._character_embedding_duplication_count = contextual_dim // character_dim
-
-        for param in self._lm.parameters():
-            param.requires_grad = requires_grad
-
-    def get_output_dim(self) -> int:
-        return self._lm._contextualizer.get_output_dim() # pylint: disable=protected-access
-
-    def forward(self,  # type: ignore
-                inputs: torch.Tensor) -> Dict[str, torch.Tensor]:
-        """
-        Parameters
-        ----------
-        inputs: ``torch.Tensor``
-            Shape ``(batch_size, timesteps, ...)`` of token ids representing the current batch.
-            These must have been produced using the same indexer the LM was trained on.
-
-        Returns
-        -------
-        The bidirectional language model representations for the input sequence, shape
-        ``(batch_size, timesteps, embedding_dim)``
-        """
-        # pylint: disable=arguments-differ
-        if self._bos_indices is not None:
-            mask = get_text_field_mask({"": inputs})
-            inputs, mask = add_sentence_boundary_token_ids(
-                    inputs, mask, self._bos_indices, self._eos_indices
-            )
-
-        source = {self._token_name: inputs}
-        result_dict = self._lm(source)
-
-        # shape (batch_size, timesteps, embedding_size)
-        noncontextual_token_embeddings = result_dict["noncontextual_token_embeddings"]
-        contextual_embeddings = result_dict["lm_embeddings"]
-
-        # Typically the non-contextual embeddings are smaller than the contextualized embeddings.
-        # Since we're averaging all the layers we need to make their dimensions match. Simply
-        # repeating the non-contextual embeddings is a crude, but effective, way to do this.
-        duplicated_character_embeddings = torch.cat(
-                [noncontextual_token_embeddings] * self._character_embedding_duplication_count, -1
-        )
-        averaged_embeddings = self._scalar_mix(
-                [duplicated_character_embeddings] + contextual_embeddings
-        )
-
-        # Add dropout
-        averaged_embeddings = self._dropout(averaged_embeddings)
-        if self._remove_bos_eos:
-            averaged_embeddings, _ = remove_sentence_boundaries(
-                    averaged_embeddings, result_dict["mask"]
-            )
-
-        return averaged_embeddings
+        super().__init__(archive_file=archive_file,
+                         dropout=dropout,
+                         bos_eos_tokens=bos_eos_tokens,
+                         remove_bos_eos=remove_bos_eos,
+                         requires_grad=requires_grad)
