+++ allennlp/tests/fixtures/language_model/bidirectional_lm_characters_token_embedder.jsonnet
@@ -1,0 +1,13 @@
+local config = import "characters_token_embedder.json";
+
+config + {
+  "model"+: {
+    "text_field_embedder"+: {
+      "token_embedders"+: {
+        "elmo"+: {
+          "type": "bidirectional_lm_token_embedder",
+        }
+      }
+    }
+  }
+}
