+++ allennlp/tests/fixtures/language_model/bidirectional_lm_characters_token_embedder_without_bos_eos.jsonnet
@@ -1,0 +1,14 @@
+local config = import "bidirectional_lm_characters_token_embedder.jsonnet";
+
+config + {
+  "model"+: {
+    "text_field_embedder"+: {
+      "token_embedders"+: {
+        "elmo"+: {
+          "bos_eos_tokens": null,
+          "remove_bos_eos": false
+        }
+      }
+    }
+  }
+}
