+++ allennlp/tests/fixtures/language_model/characters_token_embedder_without_bos_eos.jsonnet
