+++ allennlp/tests/fixtures/language_model/experiment_bidirectional.jsonnet
@@ -1,0 +1,8 @@
+local config = import "experiment_bidirectional_unsampled.jsonnet";
+
+config + {
+  "model"+: {
+    "num_samples": 10,
+    "sparse_embeddings": true
+  }
+}
