+++ allennlp/tests/fixtures/language_model/experiment_unidirectional.jsonnet
@@ -1,0 +1,8 @@
+local config = import "experiment_unidirectional_unsampled.jsonnet";
+
+config + {
+  "model"+: {
+    "num_samples": 10,
+    "sparse_embeddings": true
+  }
+}
