+++ allennlp/tests/fixtures/language_model/experiment_unidirectional_unsampled.jsonnet
@@ -1,0 +1,10 @@
+local config = import "experiment_unsampled.jsonnet";
+
+config + {
+    "model"+: {
+        "bidirectional": false,
+        "contextualizer" +: {
+            "bidirectional": false
+        }
+    }
+}
