+++ allennlp/tests/models/simple_tagger_test.py
@@ -64,11 +64,11 @@ class SimpleTaggerTest(ModelTestCase):
         training_batch = next(iterator(self.instances, num_epochs=1))
         validation_batch = next(iterator(self.instances, num_epochs=1))
 
-        training_loss = trainer.batch_loss(training_batch, for_training=True).data
-        validation_loss = trainer.batch_loss(validation_batch, for_training=False).data
+        training_loss = trainer.batch_loss(training_batch, for_training=True).item()
+        validation_loss = trainer.batch_loss(validation_batch, for_training=False).item()
 
         # Training loss should have the regularization penalty, but validation loss should not.
-        assert (training_loss == validation_loss).all()
+        numpy.testing.assert_almost_equal(training_loss, validation_loss)
 
 
 class SimpleTaggerRegularizationTest(ModelTestCase):
