+++ allennlp/tests/modules/token_embedders/bidirectional_language_model_token_embedder_test.py
@@ -1,34 +1,17 @@
 # pylint: disable=no-self-use,invalid-name
-from allennlp.common.testing import ModelTestCase
-from allennlp.data.dataset import Batch
+from .language_model_token_embedder_test import TestLanguageModelTokenEmbedder
 
 
-class TestBidirectionalLanguageModelTokenEmbedder(ModelTestCase):
+class TestBidirectionalLanguageModelTokenEmbedder(TestLanguageModelTokenEmbedder):
     def setUp(self):
         super().setUp()
-        self.set_up_model(self.FIXTURES_ROOT / 'bidirectional_lm' / 'characters_token_embedder.json',
+        self.set_up_model(self.FIXTURES_ROOT / 'language_model' /
+                          'bidirectional_lm_characters_token_embedder.jsonnet',
                           self.FIXTURES_ROOT / 'data' / 'conll2003.txt')
 
-    def test_tagger_with_bidirectional_token_embedder_can_train_save_and_load(self):
-        self.ensure_model_can_train_save_and_load(self.param_file)
-
-    def test_tagger_with_bidirectional_token_embedder_forward_pass_runs_correctly(self):
-        dataset = Batch(self.instances)
-        dataset.index_instances(self.vocab)
-        training_tensors = dataset.as_tensor_dict()
-        output_dict = self.model(**training_tensors)
-        tags = output_dict['tags']
-        assert len(tags) == 2
-        assert len(tags[0]) == 7
-        assert len(tags[1]) == 7
-        for example_tags in tags:
-            for tag_id in example_tags:
-                tag = self.model.vocab.get_token_from_index(tag_id, namespace="labels")
-                assert tag in {'O', 'I-ORG', 'I-PER', 'I-LOC'}
-
-class TestBidirectionalLanguageModelTokenEmbedderWithoutBosEos(TestBidirectionalLanguageModelTokenEmbedder):
+class TestBidirectionalLanguageModelTokenEmbedderWithoutBosEos(TestLanguageModelTokenEmbedder):
     def setUp(self):
         super().setUp()
-        self.set_up_model(self.FIXTURES_ROOT / 'bidirectional_lm' /
-                          'characters_token_embedder_without_bos_eos.jsonnet',
+        self.set_up_model(self.FIXTURES_ROOT / 'language_model' /
+                          'bidirectional_lm_characters_token_embedder_without_bos_eos.jsonnet',
                           self.FIXTURES_ROOT / 'data' / 'conll2003.txt')
