--- doc/api/allennlp.models.bidirectional_lm.rst
@@ -1,7 +1,0 @@
-allennlp.models.bidirectional_lm
-================================
-
-.. automodule:: allennlp.models.bidirectional_lm
-   :members:
-   :undoc-members:
-   :show-inheritance:
