+++ doc/api/allennlp.models.language_model.rst
@@ -1,0 +1,7 @@
+allennlp.models.language_model
+==============================
+
+.. automodule:: allennlp.models.language_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
