+++ doc/api/allennlp.models.rst
@@ -12,7 +12,6 @@ allennlp.models
   allennlp.models.archival
   allennlp.models.biaffine_dependency_parser
   allennlp.models.biattentive_classification_network
-  allennlp.models.bidirectional_lm
   allennlp.models.bimpm
   allennlp.models.constituency_parser
   allennlp.models.coreference_resolution
@@ -26,5 +25,5 @@ allennlp.models
   allennlp.models.reading_comprehension
   allennlp.models.semantic_parsing
   allennlp.models.semantic_role_labeler
+  allennlp.models.language_model
   allennlp.models.simple_tagger
-
